comp_4531/pyproject.toml
[project]
name = "comp4531"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "mlflow>=2.20.0",
    "nbformat==5.10.4",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "rpds-py==0.18.1",
    "scikeras>=0.13.0",
    "scikit-learn>=1.6.0",
    "tensorflow>=2.18.0",
    "ultralytics>=8.3.75",
]
